refactor: extract magnet toggle service
